+++ mips/lib/qps_mips.m
@@ -2,8 +2,10 @@ function [x, f, eflag, output, lambda] = qps_mips(H, c, A, l, u, xmin, xmax, x0,
 %QPS_MIPS  Quadratic Program Solver based on MIPS.
 %   [X, F, EXITFLAG, OUTPUT, LAMBDA] = ...
 %       QPS_MIPS(H, C, A, L, U, XMIN, XMAX, X0, OPT)
-%   Uses the MATPOWER Interior Point Solver (MIPS) to solve the following
-%   QP (quadratic programming) problem:
+%   [X, F, EXITFLAG, OUTPUT, LAMBDA] = QPS_MIPS(PROBLEM)
+%   A wrapper function providing a MATPOWER standardized interface for using
+%   the MATPOWER Interior Point Solver (MIPS) to solve the following QP
+%   (quadratic programming) problem:
 %
 %       min 1/2 X'*H*X + C'*X
 %        X
@@ -22,42 +24,7 @@ function [x, f, eflag, output, lambda] = qps_mips(H, c, A, l, u, xmin, xmax, x0,
 %       XMIN, XMAX : optional lower and upper bounds on the
 %           X variables, defaults are -Inf and Inf, respectively.
 %       X0 : optional starting value of optimization vector X
-%       OPT : optional options structure with the following fields,
-%           all of which are also optional (default values shown in
-%           parentheses)
-%           verbose (0) - controls level of progress output displayed
-%               0 = no progress output
-%               1 = some progress output
-%               2 = verbose progress output
-%           linsolver ('') - linear system solver for solving update steps
-%               ''          = default solver (currently same as '\')
-%               '\'         = built-in \ operator
-%               'PARDISO'   = PARDISO solver (if available)
-%           feastol (1e-6) - termination tolerance for feasibility
-%               condition
-%           gradtol (1e-6) - termination tolerance for gradient
-%               condition
-%           comptol (1e-6) - termination tolerance for complementarity
-%               condition
-%           costtol (1e-6) - termination tolerance for cost condition
-%           max_it (150) - maximum number of iterations
-%           step_control (0) - set to 1 to enable step-size control
-%           sc.red_it (20) - maximum number of step-size reductions if
-%               step-control is on
-%           cost_mult (1) - cost multiplier used to scale the objective
-%               function for improved conditioning.
-%           xi (0.99995) - constant used in alpha updates*
-%           sigma (0.1) - centering parameter*
-%           z0 (1) - used to initialize slack variables*
-%           alpha_min (1e-8) - returns EXITFLAG = -1 if either alpha
-%               parameter becomes smaller than this value*
-%           rho_min (0.95) - lower bound on rho_t*
-%           rho_max (1.05) - upper bound on rho_t*
-%           mu_threshold (1e-5) - KT multipliers smaller than this value
-%               for non-binding constraints are forced to zero
-%           max_stepsize (1e10) - returns EXITFLAG = -1 if the 2-norm of
-%               the reduced Newton step exceeds this value*
-%               * see the corresponding Appendix in the manual for details
+%       OPT : optional MIPS options struct (see MIPS for details)
 %       PROBLEM : The inputs can alternatively be supplied in a single
 %           PROBLEM struct with fields corresponding to the input arguments
 %           described above: H, c, A, l, u, xmin, xmax, x0, opt
@@ -122,7 +89,7 @@ function [x, f, eflag, output, lambda] = qps_mips(H, c, A, l, u, xmin, xmax, x0,
 %   See also MIPS.
 
 %   MIPS
-%   Copyright (c) 2010-2016, Power Systems Engineering Research Center (PSERC)
+%   Copyright (c) 2010-2020, Power Systems Engineering Research Center (PSERC)
 %   by Ray Zimmerman, PSERC Cornell
 %
 %   This file is part of MIPS.